% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.plot UPDATE.R
\name{comparison.plot}
\alias{comparison.plot}
\title{Comparing a financial parameter between two companies}
\usage{
comparison.plot(x, A, B, ticker1, ticker2)
}
\arguments{
\item{x}{A \code{string} containing the variable within the financial statement that will be plotted.}

\item{A}{A \code{matrix} containing the financial statement from the fist company.}

\item{B}{A \code{matrix} containing the financial statement from the fist company.}

\item{ticker1}{A \code{string} containing the stock ticker of the company in argument A.}

\item{ticker2}{A \code{string} containing the stock ticker of the company in argument B.}
}
\value{
A \code{plot} or \code{ggplot} object.
}
\description{
Compare any parameter on a financial statement graphically. This function makes use of the package "plotly"
}
\examples{
comparison.plot('Cash Only',create.data('wmt','Balance-Sheet'),create.data('aapl','Balance-Sheet'),'wmt','aapl')
}
\author{
Ryan Voyack
}
