% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonPlot UPDATE.R
\name{comparisonPlot}
\alias{comparisonPlot}
\title{Comparing a financial parameter between two companies}
\usage{
comparisonPlot(x, ticker1, ticker2, state)
}
\arguments{
\item{x}{A \code{numeric} containing the index number of the variable within the financial statement that will be plotted. Use createData(char,state)[[2]] to observe the vector of these varaibles.}

\item{ticker1}{A \code{string} containing the stock ticker of the company in argument A.}

\item{ticker2}{A \code{string} containing the stock ticker of the company in argument B.}

\item{state}{A \code{numeric} containing an idicator of which statement the user wants to compare variables from.
1 for Income Statement, 2 for Balance sheet and 3 for statement of cash flows}
}
\value{
A \code{plot} or \code{ggplot} object.
}
\description{
Compare any parameter on a financial statement graphically. This function makes use of the package "plotly".
    This function also makes use of the "createData" function within the financialStatementPlot package.
    This version of the function is to be used practically; not in the shiny app
}
\examples{
comparisonPlot(2,'wmt','aapl',2) #plots the "Cash Only" variable from the balance sheets of Walmart and Apple
}
\author{
Ryan Voyack
}
