% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.data function.R
\name{create.data}
\alias{create.data}
\title{Import a financial statement}
\usage{
create.data(char, state)
}
\arguments{
\item{char}{A \code{string} containing the formal stock ticker of a desired company}

\item{state}{A \code{string} containing the desired financial statement of the company
must be a character in the vector, c("Income-Statement", "Balance-Sheet", "Cash-Flows")}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{visualizeData}{A matrix containing the financial statement}
     \item{variables}{A vector of all of the variables contained on the financial statement}
}
}
\description{
Import a financial statement from the web (off of marketwatch.com). This function makes use of rvest and xml2.
}
\examples{
create.data('wmt') #imports Walmart's balance sheet with data from the last 5 years
}
\author{
Ryan Voyack
}
